// src/enums/callback-id.enum.ts
export enum CallbackId {
  OPEN_ALERT_DETAILS = "open_alert_details",
  DELETE_ALERT = "delete_alert",
  DELETE_ALL_ALERTS = "delete_all_alerts",
  CANCEL_DELETE_ALERT = "cancel_delete_alert",
  CANCEL_DELETE_ALL_ALERTS = "cancel_delete_all_alerts",
  SELECT_SEARCH_RESULT = "select_search_result",
  REFRESH_SELECTED_PRICE = "refresh_selected_price",
  CURRENT_PRICE_FROM_ACTIVE_ALERT = "current_price_from_active_alert",
}